+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - Use Bract 0.4.0
   - [TODO] Use the GA version of Bract 0.4.0
 - Bump CLJS version to 1.9.671
-- Move global/stateful vars from `config` to `global` namespace
+- Bump Asphalt version to 0.6.2
+- Move global/stateful vars from `demo.todomvc.config` to `demo.todomvc.global` namespace
 - Rename namespace `demo.todomvc.config` to `demo.todomvc.keydef`
 - [TODO] Implement `version` CLI command